Adding a local repo to be used in our tests
We should not depend on pypy`s stability to run our tests. This
commit introduces a simple http server mapping to a folder that
contains the same package structure for three simple dummy pkgs;

The binaries were added to the repo. I know it is ugly but it is
still better than accessing pypi when runing functional tests.