Installing packages from wheels
This change allows the user to install packages from wheels using the
wheelhouse (or a curd) created with curdling;

Now the `CurdManager` constructor takes the curd path, so both methods
`.get()` and `.new()` won't need to receive the path anymore. There's
also a new method called `.install()`, this is the guy that uses pip to
install things from wheel;

The acceptance tests are now running under a separate virtual
environment, so the tests are free to install any python package without
breaking the main environment. Cucumber is also helping a lot, it's not
written in python, so we just didn't have any issues running it in a
temporary virtualenv;

The deps were also updated and as of this commit, curdling depends only
on `sh` and `wheel` itself!