Adding a url to download the curd as a tar file
Now the internal web server has a url that allows the user to download
the curd as a tar file. I wouldn't spend CPU time compressing this tar
file since it contains `.whl` files which are basically `.zip` files.

This process happens in memory, there's no cache implemented as of now.