Implementing the `CurdManager.retrieve()` method
This method accesses a remote cache server, finds the requested curd,
download and extract it;

To avoid setting up the whole web server exclusively for retrieving this
package, the `urlopen` call is temporarily patched with a context
manager in the `test_retrieve_remote_curd` spec, hooking it to the
`test_client.get` retriever, so when the actual code calls:

  >>> content = urllib2.urlopen(url).read()

We're actually calling:

  >>> test_client.get(url).data

I hope it's not that complicated. The test looks pretty readable though.