Basic curd server working
Now, the main script allows the user to run a server that can be used by
other curdling users in order to just download already compiled curds;

The `-s` param was added and receives one parameter that describes both
host and port number separated by a `:` char. Just like this:

  $ python -m curdling -s localhost:8001 requirements.txt

The above line will serve a curd generated from the files received from
the command line.

In the other side, when someone else needs to use that server as a cache
server, the user just needs to issue the following command:

 $ python -m curdling -r http://localhost:8001 requirements.txt