Support forwarding `--extra-index-url` opt to pip and cleanup
This commit introduces this new variable in the `CurdManager` settings
dictionary;

Since this code was not well tested, some more unit tests were written
in order to make it safer to add this new option;

The methods `CurdManager.{new,install}` don't receive the list of
requirements anymore, it only receives the `uid` of registered
resources. To register a new resource (one requirements file or more),
you can use the `CurdManager.add` method, it will return the `uid`.