Working around pip.wheel bug and minor refactoring
The `pip wheel` command exits `0` when there's remains of an old build,
even though the installation didn't proceed well. The bug was fixed but
we still need to wait for a release, so the workaround will be around
for a while (getting into the 1.4.x series though);

The minor refactoring here was finding a better place to ensure that the
directory that will host a curd exists. We had the `install_folder`
method that was just responsible for creating the folder, now we have
the `curd_path` method that returns the path string, so we have to use
it every where we reference that path, so we won't forget to create the
directory.