Replacing `index.find` by `index.get`
The last commit introduced the new way of querying the index, now we use
the `Index.get` function instead of the old one `find`;

This replace was performed project-wide and all the functions that used
to deal with package names/indexing were refactored inside of the Index
class. Helpers like `key_from_path` and `name_from_key` were just
killed; And a new methdo `Index.index` was added to take care of
paths/files to be indexed.