Curdling server
This change introduces an exciting feature, a small server that can be
used to share your local wheels with other curdling users. There's no
security layer, we're not actually checking if the server needs the
package before writing it to disk or any other kind of sanity check;

This change introduces a couple new components, like the new download
source `CurdlingSource` which just creates a URL based on the parameters
received in `-c` and the installation request (package spec);

Another cool feature is the `Service.subscribe()` method, now services
can have multiple subscribers;

The `-u` option is also supported now. With that option, the curdling
client can send packages back to the server that doesn't seem to have
it