Addressing logging problems
I was using bare print so far. Not cool. Now we count on some thing
slightly better than that. It prints out messages granularly. You can
control that with the `--log-level` parameter that goes from 0 to the
infinite and beyond;

There's also a small refactoring of the download service. The exception
handling was improved and we're now using the log system we're talking
about here.