Implementing CurdlingLocator._get_package()
Now curdling can cache data from another curdling installation. Using
distlib as our base, the `CurdlingLocator` class just had to provide a
that method to be used along with all the other features brough by
distlib;

The `/api/package` end-point was added to our internal server and the
index was changed to make it easier to manipulate files and versions
inside of it.