Blacklist setuptools
A rudimentary mechanism for blacklisting packages for not being
processed at all was added to take `setuptools` out of the install
list. Curdling depends on `setuptools` to it *must* be installed before
you get the chance to try to install anything;

That being said, I'm also doing that cause curdling is *segfaulting*
python on Travis and I think it's related to the crazy way the
`bdist_wheel` command is called, not sure though.