First steps towards a maestro component
This component will maintain the reference to the package dependency
graph. Every time a new package is requested, the `file_package` method
will file that package under its dependency, building a tree;

In the next few commits, we will introduce the concept of closing a
node, which is just the same as saying that the node we are talking
about had all its dependencies collected;

The exit case from the recursion of download -> find_deps -> download
will be finding all the packages that have no dependencies.