Fix CurdlingLocator._get_project raising exceptions
This method is not supposed to raise any exceptions. When it fails it
should only return null. Even though it can be hard to tell the caller
what happened;

We know that instead of a 404 error we might also get an unauthorized
error and wont be able to tell the user that his credentials are just
wrong.