Workaround distlib bug fixing SimpleLocator._get_project. Closes: #4
The `distlib.lobators.SimpleScrapingLocator` class doesn't quite honor
its name. It just didn't play well with the thread model we have in
curdling, so this patch just makes this locator actually simple, by
performing all the HTTP calls in a blocking manner;

It's not a huge problem for curdling to take longer scraping a page,
since we try to separate other stuff in other threads.