Nice UI to show installation progress
Now, the Installer also has a nice UI to display its progress. It's
basically using the same approach applied to the retrieval process. The
`InstallProgress` class was renamed to `RetrieveAndBuildProgress` and
the old name is now used to provide a UI component to the installer;

The `mark` helper function in the installer is not a decorator anymore,
but it still returns a function to be connected to our servers. It's
more flexible and instead of creating a different `mark_` for each set
I want to add to the maestro, I can just call it like this:

   >>> maestro.mark('set', 'package', data)

The `Maestro.mark` method won't call `set_data` if `data` is None.