Change curdling.maestro to reflect the current tests
In a wild moment, I decided to try to fix a bug with no proper
tests. The fix was good enough to help me to come up with a new solution
to fix the actuall issue of choosing which versions should be installed
after the download process;

This idea is documented in the issue #13. This commit reverts the
previous change in order to make the code reflect all the tests declared
in the `test_maestro` module;

Small changes were done to the test file, but nothing actually
significant besides adding the test `test_maestro_mark_installed`.