Remember all the packages that requested a given dependency
Before this change, the `Maestro.mapping` structure was not prepared to
store all the packages that requested a given dependency. In other
words, the field `dependency_of` was just a string. Now it's a list, so
every time the maestro files a package, the package that requested that
dependency will be saved;

It's definitely more info to remember and it makes the whole memory
usage increase a little bit. However, we'll definitely need this data to
know exactly which packages can't be installed due to version
incompatibilities.