Maestro.should_queue should prefer packages with version info
This is part of a series of commits that is changing the method
`Maestro.should_queue()` to take version information into account;

If two different packages request different versions of the same
dependency, both dependencies will have the chance to be downloaded and
we'll wait until we actually need to install the package to decide which
version should be used. Refs #13.