Remove the last bits of Maestro that touches the main flow
The maestro component is not responsible for anything in the main flow
anymore. Now that we know that the dependencer will always yield the
right data, we just relay on it to tell us which requirements we should
aggregate to install;

The Mapping (the old Maestro) is only present in the `load_installer()`
method of the `Install` command, merging the requirements into package
names and finding the best version among all collected packages.