Skip unnecessary requirements and fix race condition in Install.feed()
There's more than one dependencer instance running all the time. So,
more than one might try to (read from|write to) the mapping, so there's
now an explicit critical region protecting this area;

It hurts the design of the pipeline a little bit and should be fixed if
that method (Or the whole "Requirement Bucket" design) gets too complex;

This detail was noticed when an optimization in which packages we should
queue was being implemented. So, it's also in this commit;

Now, the user specifies a requirement from the command line or from a
file fed to the `-r` parameter, the mapping will remember if anyone
tries to request any other version of the same requirement and it will
be just skipped.