Comment out an optimization in Install.feed()
Skipping requirements is just dangerous if it is not well tested. We were
trying to remove versions that were previously requested. It works pretty
well in the first look but it always tends to break the synchronization

tested in Install.retrieve_and_build(). This optimization can still be
done, but we might need to implement a more robust message passing
system to avoid losing the count of packages we needed to skip