Revert "Comment out an optimization in Install.feed()"
This reverts commit 1fd2d87abfaac282a91b03f0d536958b4f602fd0.

The change didn't actually fix the problem. It's actually what was
described before, but the case is different.

The leak happens when a requirement is requested more than once and the
primary requirement is a bare requirement (with no version info). E.g.:

 * gevent            # The order here is important
 * gevent (0.13.6)

If the build process fail happens, curdling will hang.