The Install component is now an Actor
Until this commit, the Install command that puts all the services
together, was not prepared for being used concurrently, leading to a
series of errors during the update of the requirement mapping;

Being an Actor means that the Install component has an Inbox fed through
the `queue` method. The processing of each item inside of this Inbox is
processed sequentially in the `retrieve_and_build()` method. This method
is doing what the `_worker()` should do and a little bit more, meaning
that it is also controlling the requirement mapping and interrupting the
loop even without receiving the sentinel;

That is actually the trick here. The biggest deal transforming the
Install command into an Actor was leveraging the two rules it plays in
the process: being the Requirement Bucket and coordinating the other
components;

Technically speaking, the loop for updating and checking the mapping
runs every X milliseconds; when this time comes, the component might not have
any packages to process yet, so that's why we do a `queue.get_nowait()`
ignoring eventual emptiness;

An eventual refactoring is strongly suggested. Separating the code that
puts the components together and the requirement bucket will definitely
improve the overall code quality and stability of the main loop.