First usable version of the `freeze` command
This command scans all the Python files under a certain directory
and. Then it scans each file looking for imported modules. After
gathering all the module names, freeze will get the name and the
version of the package that contains that module. The required
packages must be previously installed in the environment. This
information will be used to generate a `requirements.txt` file.

To illustrate exactly how to use this command, let's imagine you're
writing an application that depends on `requests`.

The first step is to import and use the library as usual:

    import math
    import requests
    # (...)
    requests.get('http://clarete.li/curdling')

The second step is to actually install the `requests` library:

    $ curd install requests

The last, but definitely not the least step is to actually call the
`freeze` command:

    $ curd freeze > requirements.txt

After that command you'll have a perfectly valid and complete
`requirements.txt` file. You shouldn't need to change it manually.